Fixing implementation to conform to linter rules.
PS: Current rules configuration warn you if line breaks are not what is expected (\r\n). This is because I use Windows (shame on me) and my Git is configured to change line endings on check-out/check-in. If you using a different environment configuration this may be a issue and you should update the linter configuration in .eslintrc.json.